Add new current layout and update README.md
Add a new layout, G-20, replacing G-18 and improve the README.

The major problem with G-18 was the use of Mod-Tap/Tap-Hold on the thumb
cluster key. With the default Tap-Hold delay in QMK this introduced a
delay long enough to be confusing and reduce accuracy, especially when
using vim. This introduces a new layout that eliminate almost all uses
of Tap-Hold uses. The exception are MOD_LALT and KC_SFTENT. Former is
less frequently used and the latter does not seem to be affected much.